requirements.txt
SpeechRecognition>=3.8.1
pyttsx3>=2.90
Pillow>=9.0.0
psutil>=5.8.0
pulsectl>=22.0.0; sys_platform == "linux"